requirements.txt
fastapi==0.62.0
uvicorn==0.12.3
mysql-connector-python==8.0.32